baru lagi

+++ database/migrations/2000_11_07_235722_create_guru_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('guru', function (Blueprint $table) {
             $table->string('nama');
             $table->primary('nama');
-            $table->integer('nip');
+            $table->unsignedBigInteger('nip');
             $table->char('JK');
             $table->timestamps();
         });
